# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Donation Station API
 * OpenAPI spec version: 0.1.0
 */
import type { DonationItem } from './donationItem';
import type { StageCount } from './stageCount';
import type { TierCount } from './tierCount';

export interface DashboardSummary {
  totalItems: number;
  byTier: TierCount[];
  byStage: StageCount[];
  recentItems: DonationItem[];
  expiringCount: number;
}