# lib/api-zod/src/generated/types/deliveryRoute.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Donation Station API
 * OpenAPI spec version: 0.1.0
 */
import type { DeliveryRouteStatus } from './deliveryRouteStatus';

export interface DeliveryRoute {
  id: string;
  name: string;
  date: Date;
  status: DeliveryRouteStatus;
  stopCount: number;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}