# lib/api-zod/src/generated/types/deliveryRouteDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Donation Station API
 * OpenAPI spec version: 0.1.0
 */
import type { DeliveryRoute } from './deliveryRoute';
import type { RouteStopDetail } from './routeStopDetail';

export type DeliveryRouteDetail = DeliveryRoute & {
  stops: RouteStopDetail[];
};