# lib/api-zod/src/generated/types/deliveryRouteInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Donation Station API
 * OpenAPI spec version: 0.1.0
 */
import type { RouteStop } from './routeStop';

export interface DeliveryRouteInput {
  name: string;
  date: Date;
  notes?: string;
  stops?: RouteStop[];
}