# lib/api-zod/src/generated/types/deliveryRouteUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Donation Station API
 * OpenAPI spec version: 0.1.0
 */
import type { DeliveryRouteUpdateStatus } from './deliveryRouteUpdateStatus';
import type { RouteStop } from './routeStop';

export interface DeliveryRouteUpdate {
  name?: string;
  date?: Date;
  status?: DeliveryRouteUpdateStatus;
  notes?: string;
  stops?: RouteStop[];
}