# lib/api-zod/src/generated/types/donationItemDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Donation Station API
 * OpenAPI spec version: 0.1.0
 */
import type { DonationItem } from './donationItem';
import type { StageHistoryEntry } from './stageHistoryEntry';

export type DonationItemDetail = DonationItem & {
  history: StageHistoryEntry[];
};