# lib/api-zod/src/generated/types/routeStop.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Donation Station API
 * OpenAPI spec version: 0.1.0
 */

export interface RouteStop {
  itemId: string;
  stopOrder: number;
  /** @nullable */
  notes?: string | null;
}