# lib/api-zod/src/generated/types/routeStopDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Donation Station API
 * OpenAPI spec version: 0.1.0
 */
import type { DonationItem } from './donationItem';

export interface RouteStopDetail {
  itemId: string;
  stopOrder: number;
  /** @nullable */
  notes?: string | null;
  item: DonationItem;
}