# lib/api-zod/src/generated/types/stageAdvance.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Donation Station API
 * OpenAPI spec version: 0.1.0
 */
import type { StageAdvanceStage } from './stageAdvanceStage';

export interface StageAdvance {
  stage: StageAdvanceStage;
  notes?: string;
}